+++ test/fixtures/pkg-with-everything/jetpack.config.mjs
@@ -1,8 +1,10 @@
 import rspackConfigForTests from '../../helpers/rspackConfigForTests.js'
 
 export default {
-  minify: false,
-  chunkLoadRetry: true,
+  build: {
+    minify: false,
+    chunkLoadRetry: true
+  },
   rspack(config) {
     rspackConfigForTests(config)
   }